requirements.txt
Flask
g4f
python-dotenv
requests
Telethon
gunicorn
